revert(common/store): remove EntityStoreWithStrategy - Allow circular dependency via the type import
