+++ packages/common/src/lib/entity/shared/index.ts
@@ -2,7 +2,6 @@ export * from './entity.enums';
 export * from './entity.interfaces';
 export * from './entity.utils';
 export * from './store';
-export * from './store-strategy';
 export * from './state';
 export * from './watcher';
 export * from './transaction';
